chore: improve RadioGroup exports
